requirements.txt
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
aiohttp==3.9.1
brotli==1.0.9
APScheduler==3.10.4
python-telegram-bot==20.7
